pyproject.toml
[project]
name = "chatbot-task"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirdon",email = "ondrej.lopuch@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "anthropic (>=0.79.0,<0.80.0)",
    "duckdb (>=1.4.4,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
